pinterest-clone/server/src/requirements.txt
Cerberus==1.3.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cloudinary==1.29.0
Flask==2.1.2
Flask-Cors==3.0.10
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
